+++ example/lib/main.dart
@@ -1,5 +1,6 @@
 import 'package:example/button_section.dart';
 import 'package:example/color_section.dart';
+import 'package:example/icon_shape_section.dart';
 import 'package:example/typography_section.dart';
 import 'package:flowbite_ui/flowbite_ui.dart';
 import 'package:flutter/material.dart';
@@ -22,7 +23,12 @@ class MainApp extends StatelessWidget {
           padding: const EdgeInsets.all(16.0),
           child: Column(
             spacing: 24.0,
-            children: [ButtonSection(), TypographySection(), ColorSection()],
+            children: [
+              IconShapeSection(),
+              ButtonSection(),
+              TypographySection(),
+              ColorSection(),
+            ],
           ),
         ),
       ),