+++ lib/src/components/components.dart
@@ -1,2 +1,3 @@
 export 'flowbite_button.dart';
+export 'flowbite_icon_shape.dart';
 export 'flowbite_outlined_button.dart';